backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Anthropic SDK
anthropic==0.42.0

# CORS and HTTP
python-multipart==0.0.6
httpx==0.26.0

# Rate limiting
slowapi==0.1.9

# Environment
python-dotenv==1.0.0

# Development
black==24.1.0
ruff==0.1.14
pytest==7.4.4
pytest-asyncio==0.23.3
